# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Скалодром «Вертикаль» — API клиентского приложения
  description: |
    REST API для мобильного приложения скалодрома «Вертикаль».

    **Источник истины** — бэкенд скалодрома. Приложение потребляет данные и инициирует
    изменения согласно use cases и entity-models из `01-analysis/04-design`.

    ### Операции, инициируемые приложением (MVP)
    - Регистрация клиента, подтверждение согласия на риск
    - Создание, отмена записей, изменение проката
    - Управление настройками уведомлений
    - Регистрация push-токена устройства

    ### Post-MVP
    - Оценка инструктора (`/ratings`)

    ### Бизнес-правила (ключевые)
    - Запись не позднее чем за 30 мин до начала (BR-006)
    - Отмена запрещена менее чем за 1 час (BR-012)
    - Допуск инструктора обязателен для «трасс с верёвкой» (BR-007)
    - Горизонт расписания — 7 дней (BR-027)
  version: 1.0.0
  contact:
    name: Скалодром «Вертикаль»
servers:
  - url: https://api.vertical-climbing.ru/v1
    description: Production
  - url: https://staging-api.vertical-climbing.ru/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
...
  /slots:
    get:
      tags: [Slots]
      summary: Получить расписание слотов
      description: |
        Список слотов тренировок за период (UC-001, FR-001, FR-002, BR-027).
        По умолчанию — 7 дней вперёд от `from`. Включает отменённые слоты (FR-006, BR-019).
      operationId: listSlots
      security: []
      parameters:
        - $ref: '#/components/parameters/DateFrom'
        - $ref: '#/components/parameters/DateTo'
        - name: date
          in: query
          description: Фильтр по конкретной дате (FR-002). Альтернатива паре from/to для одного дня.
          schema:
            type: string
            format: date
          example: "2026-07-10"
      responses:
        '200':
          description: Список слотов
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrainingSlotSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
  /slots/alternatives:
    get:
      tags: [Slots]
      summary: Найти альтернативный слот
      description: |
        Ближайший свободный слот с теми же параметрами при отмене скалодромом
        (UC-004, FR-022, BR-020, SCR-009).
      operationId: findAlternativeSlot
      parameters:
        - name: cancelled_slot_id
          in: query
          required: true
          description: ID отменённого слота
          schema:
            type: string
            format: uuid
        - name: booking_id
          in: query
          description: ID отменённой записи (для учёта параметров проката)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Альтернативный слот найден или отсутствует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlternativeSlotResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Отменённый слот или запись не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /slots/{slotId}:
    get:
      tags: [Slots]
      summary: Получить детали слота
      description: |
        Полная информация о слоте для экрана деталей (SCR-004, FR-003).
        Включает прокатный фонд и флаги доступности записи.
      operationId: getSlot
      security: []
      parameters:
        - $ref: '#/components/parameters/SlotId'
      responses:
        '200':
          description: Детали слота
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrainingSlotDetail'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Токен, полученный при регистрации (`POST /clients`)
  parameters:
    SlotId:
      name: slotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DateFrom:
      name: from
      in: query
      description: Начало периода (включительно). По умолчанию — сегодня.
      schema:
        type: string
        format: date
      example: "2026-07-04"
    DateTo:
      name: to
      in: query
      description: Конец периода (включительно). По умолчанию — from + 7 дней.
      schema:
        type: string
        format: date
      example: "2026-07-11"
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Требуется авторизация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Требуется авторизация
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Ресурс не найден
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Машиночитаемый код ошибки
        message:
          type: string
          description: Понятное сообщение для пользователя (NFR-003)
        details:
          type: object
          additionalProperties: true
          description: Дополнительные данные об ошибке
...
    SlotStatus:
      type: string
      enum: [active, cancelled_by_gym]
      description: |
        - `active` — слот доступен для записи (с учётом мест и времени)
        - `cancelled_by_gym` — отменён скалодромом, остаётся в списке (BR-019)
    FormatType:
      type: string
      enum: [bouldering_instruction, rope_routes]
      description: |
        - `bouldering_instruction` — болдеринг с инструктажем (новички)
        - `rope_routes` — трассы с верёвкой (опытные, нужен допуск)
    Difficulty:
      type: string
      enum: [beginner, experienced]
    TrainingZone:
      type: object
      required: [id, name, format_type, difficulty, max_group_size]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Болдеринг с инструктажем"
        format_type:
          $ref: '#/components/schemas/FormatType'
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        max_group_size:
          type: integer
          minimum: 1
          example: 8
    Instructor:
      type: object
      required: [id, full_name]
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
          example: "Петров Алексей Сергеевич"
        average_rating:
          type: number
          format: decimal
          minimum: 1
          maximum: 5
          nullable: true
          description: Средняя оценка 1–5 (Post-MVP / опционально MVP, FR-004)
    GymVenue:
      type: object
      required: [id, name, address]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Вертикаль"
        address:
          type: string
          example: "г. Москва, ул. Скалолазная, д. 1"
    BookingAvailability:
      type: object
      description: Вычисляемые флаги доступности записи для UI (FR-007, FR-008, FR-009)
      required:
        - can_book
        - has_free_spots
        - within_booking_window
        - clearance_required
        - clearance_granted
      properties:
        can_book:
          type: boolean
          description: Можно ли записаться (итоговый флаг)
        has_free_spots:
          type: boolean
        free_spots:
          type: integer
          minimum: 0
        within_booking_window:
          type: boolean
          description: До начала ≥ booking_cutoff_minutes
        clearance_required:
          type: boolean
          description: Требуется допуск (rope_routes)
        clearance_granted:
          type: boolean
          description: Допуск есть (актуально только если clearance_required)
    TrainingSlotSummary:
      type: object
      required:
        - id
        - starts_at
        - duration_minutes
        - capacity
        - free_spots
        - training_price
        - slot_status
        - address
        - zone
        - instructor
        - availability
      properties:
        id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
        duration_minutes:
          type: integer
          example: 90
          description: Длительность ~1,5 часа
        capacity:
          type: integer
          description: Вместимость группы (Y)
        free_spots:
          type: integer
          description: Свободные места (X)
        training_price:
          type: number
          format: decimal
        rental_tariff:
          type: number
          format: decimal
          nullable: true
          description: Базовый прокатный тариф на слот
        slot_status:
          $ref: '#/components/schemas/SlotStatus'
        address:
          type: string
        zone:
          $ref: '#/components/schemas/TrainingZone'
        instructor:
          $ref: '#/components/schemas/Instructor'
        venue:
          $ref: '#/components/schemas/GymVenue'
        availability:
          $ref: '#/components/schemas/BookingAvailability'
    TrainingSlotDetail:
      allOf:
        - $ref: '#/components/schemas/TrainingSlotSummary'
        - type: object
          required: [rental_availability]
          properties:
            rental_availability:
              type: array
              items:
                $ref: '#/components/schemas/SlotRentalAvailability'
    AlternativeSlotResponse:
      type: object
      required: [found]
      properties:
        found:
          type: boolean
          description: Найден ли альтернативный слот (BR-020)
        alternative_slot:
          $ref: '#/components/schemas/TrainingSlotSummary'
          nullable: true
          description: Ближайший свободный слот с теми же параметрами
...
    RentalEquipmentCode:
      type: string
      enum: [shoes, harness, helmet, chalk]
    RentalEquipmentType:
      type: object
      required: [id, code, name, default_price]
      properties:
        id:
          type: string
          format: uuid
        code:
          $ref: '#/components/schemas/RentalEquipmentCode'
        name:
          type: string
          example: "Скальные туфли"
        default_price:
          type: number
          format: decimal
    SlotRentalAvailability:
      type: object
      required: [id, slot_id, equipment_type_id, available_quantity, equipment_type]
      properties:
        id:
          type: string
          format: uuid
        slot_id:
          type: string
          format: uuid
        equipment_type_id:
          type: string
          format: uuid
        available_quantity:
          type: integer
          minimum: 0
        equipment_type:
          $ref: '#/components/schemas/RentalEquipmentType'
...